requirements.txt
Django==3.2.3
pytz>=2019.3
setuptools>=41.6.0
sqlparse>=0.3.0
wheel>=0.33
djangorestframework>=3.10.3
python-decouple>=3.4